removed validateIpConfigs, removed failureHandler, removed swiftv2sf_test file, removed addInterfaces config field
